+++ src/main/java/sopt/noonnu/userfont/dto/response/FontPreviewListResponse.java
@@ -1,6 +1,7 @@
-package sopt.noonnu.font.dto.response;
+package sopt.noonnu.userfont.dto.response;
 
 import sopt.noonnu.font.domain.Font;
+import sopt.noonnu.font.domain.FontMetadata;
 
 import java.util.List;
 
@@ -9,14 +10,15 @@ public record FontPreviewListResponse(
 ) {
     public static FontPreviewListResponse from(List<Font> fonts) {
         List<Item> items = fonts.stream()
-                .map(f -> new Item(f.getId(), f.getName()))
+                .map(f -> new Item(f.getId(), f.getName(), f.getFontMetadata()))
                 .toList();
         return new FontPreviewListResponse(items);
     }
 
     public record Item(
             Long id,
-            String name
+            String name,
+            FontMetadata fontMetadata
     ) {
     }
 }
